add create_RELEASES
release_handler:which_releases() will list an empty apps list
for your first release if this file is not created. If it is not
created and you do a release upgrade (for example to version "2")
then revert back to version 1 and then call
release_handler:remove_release("2") all of your apps common
between "1" and "2" will be deleted.